Rename uploadcmd to uploadui
